Refactor WordPress setup script for improved reliability and flexibility
